webapi/webapi-sensors-w3c-tests/sensors/Magnetometer_stop.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Magnetometer Test: stop()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://w3c.github.io/magnetometer/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

let sensor, stop_return;

setup(() => {
  sensor = new Magnetometer();
  sensor.start();
  stop_return = sensor.stop();
});

test(() => {
  assert_equals(sensor.state, "idle");
}, "The sensor.state changes to 'idle' after sensor.stop()");

test(() => {
  assert_equals(sensor.reading, null);
}, "the sensor.reading is null after executing stop() method");

test(() => {
  assert_throws("InvalidStateError", () => { sensor.stop(); }, "stop() twice");
}, "throw an InvalidStateError exception when state is either idle or errored");

test(() => {
  assert_equals(stop_return, undefined);
}, "the sensor.stop() returns undefined");

</script>
